try to fix tests
